script for automated release

+++ source/clearsource.bat
@@ -1,9 +1,8 @@
 @echo off
 attrib +h "*.bat" /s
 attrib +h "*.dtx" /s
-attrib +h "*.ins" /s
 attrib +h "examples\*.*" /s
 del "*.*" /s /q> nul
-rmdir svg-extract
-rmdir svg-inkscape
+if exist svg-inkscape rmdir svg-inkscape /s /q > nul
+if exist svg-extract rmdir svg-extract /s /q > nul
 attrib -h "*.*" /s
